requirements.txt
django-5.2.7